requirements.txt
pyyaml==6.0.2
numpy<2.0
